module-adapter: remove .cmd() from IPC4
struct comp_ops::cmd() is only used for IPC3, remove it and
comp_cmd() from the IPC4 version.
