+++ src/audio/module_adapter/module_adapter_ipc3.c
@@ -165,3 +165,125 @@ int module_adapter_set_state(struct processing_module *mod, struct comp_dev *dev
 	return comp_set_state(dev, cmd);
 }
 
+static int module_adapter_get_set_params(struct comp_dev *dev, struct sof_ipc_ctrl_data *cdata,
+					 bool set)
+{
+	struct processing_module *mod = comp_get_drvdata(dev);
+	struct module_data *md = &mod->priv;
+	enum module_cfg_fragment_position pos;
+	uint32_t data_offset_size;
+	static uint32_t size;
+
+	comp_dbg(dev, "module_adapter_set_params(): num_of_elem %d, elem remain %d msg_index %u",
+		 cdata->num_elems, cdata->elems_remaining, cdata->msg_index);
+
+	/* set the fragment position, data offset and config data size */
+	if (!cdata->msg_index) {
+		size = cdata->num_elems + cdata->elems_remaining;
+		data_offset_size = size;
+		if (cdata->elems_remaining)
+			pos = MODULE_CFG_FRAGMENT_FIRST;
+		else
+			pos = MODULE_CFG_FRAGMENT_SINGLE;
+	} else {
+		data_offset_size = size - (cdata->num_elems + cdata->elems_remaining);
+		if (cdata->elems_remaining)
+			pos = MODULE_CFG_FRAGMENT_MIDDLE;
+		else
+			pos = MODULE_CFG_FRAGMENT_LAST;
+	}
+
+	/*
+	 * The type member in struct sof_abi_hdr is used for component's specific blob type
+	 * for IPC3, just like it is used for component's specific blob param_id for IPC4.
+	 */
+	if (set && md->ops->set_configuration)
+		return md->ops->set_configuration(mod, cdata->data[0].type, pos, data_offset_size,
+						  (const uint8_t *)cdata, cdata->num_elems,
+						  NULL, 0);
+	else if (!set && md->ops->get_configuration)
+		return md->ops->get_configuration(mod, pos, &data_offset_size,
+						  (uint8_t *)cdata, cdata->num_elems);
+
+	comp_warn(dev, "module_adapter_get_set_params(): no configuration op set for %d",
+		  dev_comp_id(dev));
+	return 0;
+}
+
+static int module_adapter_ctrl_get_set_data(struct comp_dev *dev, struct sof_ipc_ctrl_data *cdata,
+					    bool set)
+{
+	int ret;
+	struct processing_module *mod = comp_get_drvdata(dev);
+
+	comp_dbg(dev, "module_adapter_ctrl_set_data() start, state %d, cmd %d",
+		 mod->priv.state, cdata->cmd);
+
+	/* Check version from ABI header */
+	if (SOF_ABI_VERSION_INCOMPATIBLE(SOF_ABI_VERSION, cdata->data->abi)) {
+		comp_err(dev, "module_adapter_ctrl_set_data(): ABI mismatch!");
+		return -EINVAL;
+	}
+
+	switch (cdata->cmd) {
+	case SOF_CTRL_CMD_ENUM:
+		comp_err(dev, "module_adapter_ctrl_set_data(): set enum is not implemented");
+		ret = -EIO;
+		break;
+	case SOF_CTRL_CMD_BINARY:
+		ret = module_adapter_get_set_params(dev, cdata, set);
+		break;
+	default:
+		comp_err(dev, "module_adapter_ctrl_set_data error: unknown set data command");
+		ret = -EINVAL;
+		break;
+	}
+
+	return ret;
+}
+
+/* Used to pass standard and bespoke commands (with data) to component */
+int module_adapter_cmd(struct comp_dev *dev, int cmd, void *data, int max_data_size)
+{
+	struct sof_ipc_ctrl_data *cdata = ASSUME_ALIGNED(data, 4);
+	struct processing_module *mod = comp_get_drvdata(dev);
+	struct module_data *md = &mod->priv;
+	int ret = 0;
+
+	comp_dbg(dev, "module_adapter_cmd() %d start", cmd);
+
+	switch (cmd) {
+	case COMP_CMD_SET_DATA:
+		ret = module_adapter_ctrl_get_set_data(dev, cdata, true);
+		break;
+	case COMP_CMD_GET_DATA:
+		ret = module_adapter_ctrl_get_set_data(dev, cdata, false);
+		break;
+	case COMP_CMD_SET_VALUE:
+		/*
+		 * IPC3 does not use config_id, so pass 0 for config ID as it will be ignored
+		 * anyway. Also, pass the 0 as the fragment size as it is not relevant for the
+		 * SET_VALUE command.
+		 */
+		if (md->ops->set_configuration)
+			ret = md->ops->set_configuration(mod, 0, MODULE_CFG_FRAGMENT_SINGLE, 0,
+							  (const uint8_t *)cdata, 0, NULL, 0);
+		break;
+	case COMP_CMD_GET_VALUE:
+		/*
+		 * IPC3 does not use config_id, so pass 0 for config ID as it will be ignored
+		 * anyway. Also, pass the 0 as the fragment size and data offset as they are not
+		 * relevant for the GET_VALUE command.
+		 */
+		if (md->ops->get_configuration)
+			ret = md->ops->get_configuration(mod, 0, 0, (uint8_t *)cdata, 0);
+		break;
+	default:
+		comp_err(dev, "module_adapter_cmd() error: unknown command");
+		ret = -EINVAL;
+		break;
+	}
+
+	comp_dbg(dev, "module_adapter_cmd() done");
+	return ret;
+}
